simplified the user model
removed the array items from the user. items of the user are now loaded
from the Item class. to buy an item it is sufficient to change the
owner on the item (and push some money around of course). you don't
have to push it to the item list of the buyer and remove it from the
owner.